added separated views in app files
