+++ app/db/mocked_data.py
@@ -59,9 +59,9 @@ CANDIDATES = [
         'address': 'Boeiro de Rua, 1, New York, NY, USA',
         'phone': '+1 (212) 846-2543',
         'linkedin': 'www.linkedin.com/michelangelo',
-        'Applied_On': 'December 10, 1815',
-        'Stage': 'Applied',
-        'last_modify_on': 'December 10, 1815',
+        'applied_on': 'December 10, 1815',
+        'stage': 'Applied',
+        'last_modified_on': 'December 10, 1815',
         'email': 'michelangelo@tmnt.com',
     },
     {
@@ -70,9 +70,9 @@ CANDIDATES = [
         'address': 'Boeiro de Rua, 1, New York, NY, USA',
         'phone': '+1 (212) 846-2543',
         'linkedin': 'www.linkedin.com/leonardo',
-        'Applied_On': 'December 10, 1815',
-        'Stage': 'Contacted',
-        'last_modify_on': 'December 10, 1815',
+        'applied_on': 'December 10, 1815',
+        'stage': 'Contacted',
+        'last_modified_on': 'December 10, 1815',
         'email': 'leonardo@tmnt.com',
     },
     {
@@ -81,9 +81,9 @@ CANDIDATES = [
         'address': 'Boeiro de Rua, 1, New York, NY, USA',
         'phone': '+1 (212) 846-2543',
         'linkedin': 'www.linkedin.com/rafael',
-        'Applied_On': 'December 10, 1815',
-        'Stage': 'Resume Analysis',
-        'last_modify_on': 'December 10, 1815',
+        'applied_on': 'December 10, 1815',
+        'stage': 'Resume Analysis',
+        'last_modified_on': 'December 10, 1815',
         'email': 'rafael@tmnt.com',
     },
     {
@@ -92,9 +92,9 @@ CANDIDATES = [
         'address': 'Boeiro de Rua, 1, New York, NY, USA',
         'phone': '+1 (212) 846-2543',
         'linkedin': 'www.linkedin.com/donatello',
-        'Applied_On': 'December 10, 1815',
-        'Stage': 'Interview Done',
-        'last_modify_on': 'December 10, 1815',
+        'applied_on': 'December 10, 1815',
+        'stage': 'Interview Done',
+        'last_modified_on': 'December 10, 1815',
         'email': 'donatello@tmnt.com',
     },
     {
@@ -103,9 +103,9 @@ CANDIDATES = [
         'address': 'Boeiro de Rua, 1, New York, NY, USA',
         'phone': '+1 (212) 846-2543',
         'linkedin': 'www.linkedin.com/splinter',
-        'Applied_On': 'December 10, 1815',
-        'Stage': 'Applied',
-        'last_modify_on': 'December 10, 1815',
+        'applied_on': 'December 10, 1815',
+        'stage': 'Applied',
+        'last_modified_on': 'December 10, 1815',
         'email': 'splinter@tmnt.com',
     },
 ]
