+++ app/db/mongodb/candidates.py
@@ -5,13 +5,13 @@ def find_all ():
     result = list(DATABASE.candidates.find({}, {'_id': False}))
     return result
 
-def find_by_stage(Stage):
-    result = list(DATABASE.candidates.find({'Stage': Stage}, {'_id': False}))
+def find_by_stage(stage):
+    result = list(DATABASE.candidates.find({'stage': stage}, {'_id': False}))
     return result
 
 def find_one(candidate_id):
     result = DATABASE.candidates.find_one({'id': candidate_id}, {'_id': False})
     return result
 
-def move_candidate(candidate_id, Stage):
-    DATABASE.candidates.update_one({'id': candidate_id}, {'$set': {'Stage': Stage}})
+def move_candidate(candidate_id, stage):
+    DATABASE.candidates.update_one({'id': candidate_id}, {'$set': {'stage': stage}})
